(~) Pin-Override-Helfer in OverridePinMapper extrahieren (#561, fixt FileSizeLimit)
